+++ bootstrap.php
@@ -2,5 +2,4 @@
 
 require_once __DIR__ . '/vendor/autoload.php';
 
-define('OMEKACLI_PATH', __DIR__ . '/..');
 define('OMEKACLI_VERSION', '1.0.0-alpha.1');