+++ src/Command/CheckUpdatesCommand.php
@@ -2,8 +2,10 @@
 
 namespace OmekaCli\Command;
 
+use Zend_Registry;
 use OmekaCli\Application;
 use OmekaCli\Command\PluginCommands\Update;
+use OmekaCli\Plugin\Updater;
 
 class CheckUpdatesCommand extends AbstractCommand
 {
@@ -25,47 +27,35 @@ class CheckUpdatesCommand extends AbstractCommand
 
     public function run($options, $args, Application $application)
     {
-        if (file_exists(OMEKACLI_PATH . '/.git')) {
-            $output = shell_exec('git -C ' . OMEKACLI_PATH . ' log --oneline HEAD..@{u}');
-            if (empty($output)) {
-                $this->logger->info('omeka-cli is up-to-date.');
-            } else {
-                echo 'omeka-cli' . PHP_EOL;
-            }
+        # TODO Use GitHub releases API
+        $remoteTag = rtrim(`git ls-remote -q --tags --refs https://github.com/biblibre/omeka-cli | cut -f 2 | sed "s|refs/tags/||" | sort -rV | head -n1`);
+        $remoteVersion = ltrim($remoteTag, 'v');
+        if (OMEKACLI_VERSION === $remoteVersion) {
+            $this->logger->info('omeka-cli is up-to-date ({version})', array('version' => OMEKACLI_VERSION));
         } else {
-            $remoteVersion = shell_exec('git -C ' . OMEKACLI_PATH . ' ls-remote --tags https://github.com/biblibre/omeka-cli 2>/dev/null | grep -o \'[0-9]\+\.[0-9]\+\.[0-9]\+\' | sort -rV | sed 1q');
-            if (OMEKACLI_VERSION == $remoteVersion) {
-                $this->logger->info('omeka-cli is up-to-date.');
+            echo sprintf('omeka-cli (%s -> %s)', OMEKACLI_VERSION, $remoteVersion) . PHP_EOL;
+        }
+
+        if ($application->isOmekaInitialized()) {
+            $latestOmekaVersion = latest_omeka_version();
+            if (version_compare(OMEKA_VERSION, $latestOmekaVersion) >= 0) {
+                $this->logger->info('Omeka is up-to-date ({version})', array('version' => OMEKA_VERSION));
             } else {
-                echo 'omeka-cli' . PHP_EOL;
+                echo sprintf('Omeka (%s -> %s)', OMEKA_VERSION, $latestOmekaVersion) . PHP_EOL;
+            }
+
+            $updater = new Updater();
+            $updater->setLogger($this->logger);
+            $pluginLoader = Zend_Registry::get('plugin_loader');
+            $plugins = $pluginLoader->getPlugins();
+            foreach ($plugins as $plugin) {
+                $latestVersion = $updater->getPluginLatestVersion($plugin);
+                if (version_compare($latestVersion, $plugin->getIniVersion()) > 0) {
+                    echo sprintf('%s (%s -> %s)', $plugin->name, $plugin->getIniVersion(), $latestVersion) . PHP_EOL;
+                } else {
+                    $this->logger->info('{plugin} is up-to-date ({version})', array('plugin' => $plugin->name, 'version' => $plugin->getIniVersion()));
+                }
             }
         }
-
-        if (!$application->isOmekaInitialized()) {
-            $this->logger->error('Omeka is not initialized here.');
-
-            return 1;
-        }
-
-        $db = get_db();
-        $pluginsTable = $db->getTable('Plugin');
-        $activePlugins = $pluginsTable->findBy(array('active' => 1));
-        $inactivePlugins = $pluginsTable->findBy(array('active' => 0));
-
-        if (version_compare(OMEKA_VERSION, latest_omeka_version()) >= 0) {
-            $this->logger->info('Omeka is up-to-date.');
-        } else {
-            echo 'Omeka' . PHP_EOL;
-        }
-
-        if (OMEKA_VERSION != get_option('omeka_version')) {
-            $this->logger->warning('Omeka version and database version do not match!');
-        }
-
-        $this->logger->info('Plugins status:');
-        $updateCommand = new Update();
-        $updateCommand->setLogger($this->logger);
-
-        return $updateCommand->run(array('list' => true), array(), $application);
     }
 }