+++ src/Command/PluginCommands/Install.php
@@ -2,11 +2,11 @@
 
 namespace OmekaCli\Command\PluginCommands;
 
+use Zend_Registry;
+use Plugin;
 use OmekaCli\Application;
-use OmekaCli\Command\AbstractCommand;
-use OmekaCli\Command\PluginCommands\Utils\PluginUtils as PUtils;
 
-class Install extends AbstractCommand
+class Install extends AbstractPluginCommand
 {
     public function getDescription()
     {
@@ -15,66 +15,53 @@ class Install extends AbstractCommand
 
     public function getUsage()
     {
-        return 'usage:' . PHP_EOL
-             . '    plugin-install PLUGIN_NAME' . PHP_EOL
-             . '    plin PLUGIN_NAME' . PHP_EOL
-             . PHP_EOL
-             . 'Install a plugin' . PHP_EOL;
+        return "Usage:\n"
+             . "\tplugin-install PLUGIN_NAME\n"
+             . "\tplin PLUGIN_NAME\n";
     }
 
     public function run($options, $args, Application $application)
     {
-        if (count($args) == 1) {
-            $pluginName = array_shift($args);
-        } else {
-            $this->logger->error($this->getUsage());
-
-            return 1;
-        }
-
-        $this->logger->info('Checking Omeka status');
         if (!$application->isOmekaInitialized()) {
-            $this->logger->error('omeka not initialized here');
+            $this->logger->error('Omeka is not initialized here');
 
             return 1;
         }
 
-        $this->logger->info('Checking plugin directory');
-        if (!file_exists(PLUGIN_DIR . '/' . $pluginName)) {
-            $this->logger->error('plugin not found');
+        if (count($args) != 1) {
+            $this->logger->error('Bad number of arguments');
+            error_log($this->getUsage());
 
             return 1;
         }
 
-        $this->logger->info('Checking dependencies');
-        $missingDeps = PUtils::getMissingDependencies($pluginName);
-        if (!empty($missingDeps)) {
-            $this->logger->error('error: missing plugins ' . implode(',', $missingDeps));
+        $pluginName = reset($args);
 
+        $plugin = $this->getPlugin($pluginName);
+        if ($plugin) {
+            $this->logger->error('{plugin} is already installed', array('plugin' => $plugin->name));
             return 1;
         }
 
-        $this->logger->info('Creating new plugin');
-        $ini = parse_ini_file(PLUGIN_DIR . '/' . $pluginName . '/plugin.ini');
-        $version = $ini['version'];
-
-        $plugin = new \Plugin();
+        $plugin = new Plugin();
         $plugin->name = $pluginName;
-        $plugin->setIniVersion($version);
-        $plugin->setLoaded(true);
 
-        $this->logger->info('Installing plugin');
-        $installer = PUtils::getInstaller($plugin);
-        $installer->install($plugin);
+        $this->getPluginIniReader()->load($plugin);
 
-        $results = get_db()->getTable('Plugin')->findBy(array(
-            'active' => 1,
-            'name' => $plugin->name,
-        ));
-        if (!empty($results) && array_shift($results)->isActive()) {
-            $this->logger->info('Installation succeeded');
+        $missingDeps = $this->getMissingDependencies($plugin);
+        if (!empty($missingDeps)) {
+            $this->logger->error('Some required plugins are missing: ' . implode(', ', $missingDeps));
+
+            return 1;
+        }
+
+        $this->getPluginInstaller()->install($plugin);
+
+        $plugin = get_db()->getTable('Plugin')->findByDirectoryName($plugin->name);
+        if ($plugin && $plugin->isActive()) {
+            $this->logger->info('{plugin} installed', array('plugin' => $plugin->name));
         } else {
-            $this->logger->error('installation failed');
+            $this->logger->error('Installation of {plugin} failed', array('plugin' => $plugin->name));
         }
 
         return 0;