+++ src/Command/PluginCommands/Update.php
@@ -2,21 +2,13 @@
 
 namespace OmekaCli\Command\PluginCommands;
 
+use Omeka_Plugin_Ini;
+use Omeka_Plugin_Installer_Exception;
 use OmekaCli\Application;
-use OmekaCli\Command\AbstractCommand;
-use GetOptionKit\OptionCollection;
-use Github\Client;
+use OmekaCli\Plugin\Updater;
 
-class Update extends AbstractCommand
+class Update extends AbstractPluginCommand
 {
-    public function getOptionsSpec()
-    {
-        $cmdSpec = new OptionCollection();
-        $cmdSpec->add('list', 'list plugins to update only');
-
-        return $cmdSpec;
-    }
-
     public function getDescription()
     {
         return 'update plugins';
@@ -24,110 +16,78 @@ class Update extends AbstractCommand
 
     public function getUsage()
     {
-        return 'usage:' . PHP_EOL
-             . '    plugin-update [--list] [PLUGIN_NAME]' . PHP_EOL
-             . '    plup [--list] [PLUGIN_NAME]' . PHP_EOL
-             . PHP_EOL
-             . 'Update plugins. Use the --list option to get the list of '
-             . 'plugins to update without updating them.' . PHP_EOL;
+        return "Usage:\n"
+             . "\tplugin-update [<plugin_name>]\n"
+             . "\n"
+             . "If <plugin_name> is given, update only this plugin.\n";
     }
 
     public function run($options, $args, Application $application)
     {
-        if (!empty($options)) {
-            $listOnly = $options['list'];
-        } else {
-            $listOnly = false;
-        }
-        switch (count($args)) {
-        case 0:
-            $pluginName = null;
-            break;
-        case 1:
-            $pluginName = array_pop($args);
-            break;
-        default:
-            $this->logger->error($this->getUsage());
-
-            return 1;
-        }
-
         if (!$application->isOmekaInitialized()) {
             $this->logger->error('Omeka not initialized here.');
 
             return 1;
         }
-        $c = new Client();
-        $plugins = get_db()->getTable('Plugin')->findBy($pluginName == null ? array() : array('name' => $pluginName));
+
+        if (count($args) > 1) {
+            $this->logger->error('Bad number of arguments');
+            error_log($this->getUsage());
+
+            return 1;
+        }
+
+        $pluginName = reset($args);
+
+        if (!empty($pluginName)) {
+            $plugin = $this->getPlugin($pluginName);
+            if (!isset($plugin)) {
+                $this->logger->error('Plugin {plugin} not found', array('plugin' => $pluginName));
+
+                return 1;
+            }
+
+            $plugins = array($plugin);
+        } else {
+            $plugins = $this->getPluginLoader()->getPlugins();
+        }
+
+        $pluginInstaller = $this->getPluginInstaller();
+        $updater = new Updater();
+        $updater->setLogger($this->logger);
+
         foreach ($plugins as $plugin) {
-            if (!$listOnly) {
-                $this->logger->info('updating ' . $plugin->name);
+            $latestVersion = $updater->getPluginLatestVersion($plugin);
+            if (version_compare($latestVersion, $plugin->getIniVersion()) <= 0) {
+                $this->logger->notice('{plugin} is up-to-date ({version})', array('plugin' => $plugin->name, 'version' => $plugin->getIniVersion()));
+                continue;
             }
-            if (file_exists(PLUGIN_DIR . '/' . $plugin->name . '/.git')) {
-                // TODO: Move github specific code to GitHub repo class.
-                system('git -C ' . PLUGIN_DIR . '/' . $plugin->name . ' rev-parse @{u} 1>/dev/null 2>/dev/null', $ans);
-                if ($ans != 0) {
-                    $this->logger->error('no upstream.');
-                    continue;
-                }
-                shell_exec('git -C ' . PLUGIN_DIR . '/' . $plugin->name . ' fetch 2>/dev/null');
-                $output = shell_exec('git -C ' . PLUGIN_DIR . '/' . $plugin->name . ' log --oneline HEAD..@{u}');
-                if (empty($output)) {
-                    $this->logger->info($plugin->name . ' is up-to-date');
-                    continue;
-                }
-                if ($listOnly) {
-                    echo $plugin->name . ' can be updated' . PHP_EOL;
-                    continue;
-                } else {
-                    $this->logger->info('new version available.');
-                }
-                shell_exec('git -C ' . PLUGIN_DIR . '/' . $plugin->name . ' pull --rebase');
-            } else {
-                $repoClass = 'OmekaCli\Command\PluginCommands\Utils\Repository\OmekaDotOrgRepository';
-                $repo = new $repoClass();
-                $version = $repo->findPlugin($plugin->name)['url'];
-                $tmp = preg_replace('/\.zip$/', '', preg_split('/-/', $version));
-                $version = end($tmp);
-                if ($plugin->version == $version) {
-                    continue;
-                } else {
-                    if ($listOnly) {
-                        echo $plugin->name . ' can be updated' . PHP_EOL;
-                        continue;
-                    }
-                    $backDir = getenv('HOME') . '/.omeka-cli/backups';
-                    if (!is_dir($backDir)) {
-                        if (!mkdir($backDir, 0755, true)) {
-                            if (!UIUtils::confirmPrompt('Cannot create backups directory. Anyway?')) {
-                                continue;
-                            }
-                        }
-                    }
-                    shell_exec('mv ' . PLUGIN_DIR . '/' . $plugin->name . ' '
-                                     . $backDir . '/' . $plugin->name . '_' . date('YmdHi'));
-                    try {
-                        $pluginInfo = $repo->find($plugin->name);
-                        $repo->download($pluginInfo, PLUGIN_DIR);
-                    } catch (\Exception $e) {
-                        $this->logger->error('cannot update plugin.');
-                        echo $e->getMessage() . PHP_EOL;
-                        shell_exec('mv ' . $backDir . '/' . $plugin->name . '_' . date('YmdHi') . ' '
-                                         . PLUGIN_DIR . '/' . $plugin->name);
-                    }
-                }
+
+            // If a plugin is active, its code is already loaded at this point
+            // and will not be reloaded after the update, so the upgrade hook
+            // will not be up-to-date. We have to force the user to manually
+            // deactivate the plugin first.
+            // TODO Find another way
+            if ($plugin->isActive()) {
+                $this->logger->error('{plugin} is active and needs to be deactivated before being updated', array('plugin' => $plugin->name));
+                continue;
             }
-            $broker = $plugin->getPluginBroker();
-            $loader = new \Omeka_Plugin_Loader($broker,
-                                              new \Omeka_Plugin_Ini(PLUGIN_DIR),
-                                              new \Omeka_Plugin_Mvc(PLUGIN_DIR),
-                                              PLUGIN_DIR);
-            $installer = new \Omeka_Plugin_Installer($broker, $loader);
-            if (null === $plugin->getIniVersion()) {
-                $version = parse_ini_file(PLUGIN_DIR . '/' . $plugin->name . '/plugin.ini')['version'];
-                $plugin->setIniVersion($version);
+
+            $this->logger->info('Updating {plugin}', array('plugin' => $plugin->name));
+            if (!$updater->update($plugin)) {
+                continue;
             }
-            $installer->upgrade($plugin);
+
+            $iniReader = new Omeka_Plugin_Ini(PLUGIN_DIR);
+            $iniReader->load($plugin);
+
+            try {
+                $iniVersion = $plugin->getIniVersion();
+                $pluginInstaller->upgrade($plugin);
+                $this->logger->info('Plugin {plugin} upgraded successfully to {version}', array('plugin' => $plugin->name, 'version' => $iniVersion));
+            } catch (Omeka_Plugin_Installer_Exception $e) {
+                $this->logger->error('Failed to upgrade {plugin}: {message}', array('plugin' => $plugin->name, 'message' => $e->getMessage()));
+            }
         }
 
         return 0;